requirements.txt
six
pyinstaller
colorama
orator
pycryptodome
PySimpleGUI
requests